move high_res_clock.h to the utility folder
