+++ cpp/tests/components/con_comp_test.cu
@@ -14,7 +14,7 @@
 
 #include "cuda_profiler_api.h"
 #include "gtest/gtest.h"
-#include "high_res_clock.h"
+#include <utilities/high_res_clock.h>
 
 #include <algorithm>
 #include <algorithms.hpp>
